feat: implement deeplink functionality and refactor QR upload process
- Added a new DeeplinkPage component to handle deeplink navigation.
- Introduced useProcessDeeplinkMutation for processing deeplink URLs.
- Updated UploadQrPage to navigate to the new deeplink feature upon QR code scanning.
- Refactored QR upload components to streamline the user experience.
- Removed deprecated QR upload functionality from the extension and desktop clients.
- Enhanced core state management to include getMpcServerUrl for improved server URL handling.
