+++ clients/desktop/src/AppProviders.tsx
@@ -1,4 +1,5 @@
 import buildInfo from '@clients/desktop/build.json'
+import { mpcServerUrl } from '@core/mpc/MpcServerType'
 import { WalletCoreProvider } from '@core/ui/chain/providers/WalletCoreProvider'
 import { VaultCreationMpcLibProvider } from '@core/ui/mpc/state/vaultCreationMpcLib'
 import { CoreProvider, CoreState } from '@core/ui/state/core'
@@ -14,6 +15,7 @@ import { QueryClientProvider } from '@tanstack/react-query'
 import { BrowserOpenURL, ClipboardGetText } from '@wailsapp/runtime'
 
 import { SaveFile } from '../wailsjs/go/main/App'
+import { DiscoveryService } from '../wailsjs/go/mediator/Server'
 import { useVaultCreationMpcLib } from './mpc/state/vaultCreationMpcLib'
 import { getQueryClient } from './query/queryClient'
 import { storage } from './state/storage'
@@ -37,6 +39,13 @@ const coreState: CoreState = {
   getClipboardText: ClipboardGetText,
   version: buildInfo.version,
   isLocalModeAvailable: true,
+  getMpcServerUrl: async ({ serverType, serviceName }) => {
+    if (serverType === 'relay') {
+      return mpcServerUrl.relay
+    }
+
+    return DiscoveryService(serviceName)
+  },
 }
 
 export const AppProviders = ({ children }: ChildrenProp) => {
