+++ clients/extension/src/providers/AppProviders.tsx
@@ -1,5 +1,6 @@
 import { QueryProvider } from '@clients/extension/src/providers/QueryClientProvider'
 import { MpcLib } from '@core/mpc/mpcLib'
+import { mpcServerUrl } from '@core/mpc/MpcServerType'
 import { WalletCoreProvider } from '@core/ui/chain/providers/WalletCoreProvider'
 import { VaultCreationMpcLibProvider } from '@core/ui/mpc/state/vaultCreationMpcLib'
 import { CoreProvider, CoreState } from '@core/ui/state/core'
@@ -39,6 +40,13 @@ const coreState: CoreState = {
   getClipboardText: () => navigator.clipboard.readText(),
   version: '1.0.0',
   isLocalModeAvailable: false,
+  getMpcServerUrl: async ({ serverType }) => {
+    if (serverType === 'relay') {
+      return mpcServerUrl.relay
+    }
+
+    throw new Error('Local mode is not available in extension')
+  },
 }
 
 export const AppProviders = ({ children }: ChildrenProp) => {
