+++ core/ui/deeplink/mutations/useProcessDeeplinkMutation.ts
@@ -4,13 +4,18 @@ import {
   tssMessageSchema,
   TssType,
 } from '@core/mpc/types/utils/tssType'
+import {
+  KeysignMessageSchema,
+  KeysignPayloadSchema,
+} from '@core/mpc/types/vultisig/keysign/v1/keysign_message_pb'
 import { useCoreNavigate } from '@core/ui/navigation/hooks/useCoreNavigate'
+import { decompressQrPayload } from '@core/ui/qr/upload/utils/decompressQrPayload'
 import { match } from '@lib/utils/match'
 import { getRawQueryParams } from '@lib/utils/query/getRawQueryParams'
 import { useMutation } from '@tanstack/react-query'
 
-import { parseTransferredKeysignMsg } from '../../vault/keysign/shared/utils/parseTransfferedKeysignMsg'
-import { decompressQrPayload } from '../../vault/qr/upload/utils/decompressQrPayload'
+import { getPayloadFromServer } from '../../mpc/keygen/create/fast/server/utils/getPayloadFromServer'
+import { useCore } from '../../state/core'
 
 type DeeplinkType = 'NewVault' | 'SignTransaction'
 
@@ -29,6 +34,8 @@ type DeeplinkParams = DeeplinkSharedData & {
 
 export const useProcessDeeplinkMutation = () => {
   const navigate = useCoreNavigate()
+
+  const { getMpcServerUrl } = useCore()
 
   return useMutation({
     mutationFn: async (url: string) => {
@@ -62,7 +69,27 @@ export const useProcessDeeplinkMutation = () => {
           SignTransaction: async () => {
             const vaultId = queryParams.vault
 
-            const keysignMsg = await parseTransferredKeysignMsg(payload)
+            const keysignMsg = fromBinary(KeysignMessageSchema, payload)
+
+            if (keysignMsg.payloadId) {
+              const serverType = keysignMsg.useVultisigRelay ? 'relay' : 'local'
+              const serverUrl = await getMpcServerUrl({
+                serverType,
+                serviceName: keysignMsg.serviceName,
+              })
+
+              const rawPayload = await getPayloadFromServer({
+                hash: keysignMsg.payloadId,
+                serverUrl,
+              })
+              const payload = await decompressQrPayload(rawPayload)
+
+              keysignMsg.payloadId = ''
+              keysignMsg.keysignPayload = fromBinary(
+                KeysignPayloadSchema,
+                payload
+              )
+            }
 
             navigate(
               {
