+++ core/ui/navigation/sharedViews.tsx
@@ -23,6 +23,9 @@ import { VaultFolderPage } from '@core/ui/vaultsOrganisation/folder/VaultFolderP
 import { CreateVaultFolderPage } from '@core/ui/vaultsOrganisation/folders/create/CreateVaultFolderPage'
 import { Views } from '@lib/ui/navigation/Views'
 
+import { DeeplinkPage } from '../deeplink/components/DeeplinkPage'
+import { UploadQrPage } from '../qr/upload/UploadQrPage'
+
 export type SharedViewId = Extract<
   CoreViewId,
   | 'airdropRegister'
@@ -46,6 +49,8 @@ export type SharedViewId = Extract<
   | 'vaultFolder'
   | 'manageVaultFolder'
   | 'deposit'
+  | 'uploadQr'
+  | 'deeplink'
 >
 
 export const sharedViews: Views<SharedViewId> = {
@@ -78,4 +83,6 @@ export const sharedViews: Views<SharedViewId> = {
       <VaultFolderPage />
     </CurrentVaultFolderPageProvider>
   ),
+  uploadQr: UploadQrPage,
+  deeplink: DeeplinkPage,
 }
