+++ core/ui/qr/upload/UploadQrPage/index.tsx
