+++ core/ui/qr/upload/UploadQrView.tsx
