+++ core/ui/qr/upload/useProcessQrMutation.ts
@@ -1,10 +1,10 @@
 import { readQrCode } from '@core/ui/qr/utils/readQrCode'
 import { useMutation } from '@tanstack/react-query'
 
-import { useAppNavigate } from '../../../navigation/hooks/useAppNavigate'
+import { useCoreNavigate } from '../../navigation/hooks/useCoreNavigate'
 
 export const useProcessQrMutation = () => {
-  const navigate = useAppNavigate()
+  const navigate = useCoreNavigate()
 
   return useMutation({
     mutationFn: async (file: File) => {
